# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "xg backend rest api",
        "title": "xg REST API",
        "termsOfService": "https://localhost:8088/v1",
        "contact": {},
        "version": "0.1 alpha"
    },
    "paths": {
...
        "/api/role": {
            "post": {
                "description": "create a new role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "createRole",
                "parameters": [
                    {
                        "description": "create role request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.CreateRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.IdResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
        "/api/roles": {
            "get": {
                "description": "list all roles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "listRoles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Role"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
        "/api/statistics/summary": {
            "get": {
                "description": "get system data graph",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "graph",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.StatisticGraph"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
        "/api/student": {
            "post": {
                "description": "create a new student",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "createStudent",
                "parameters": [
                    {
                        "description": "create student request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.CreateStudentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.IdResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
        "/api/student/{id}": {
            "get": {
                "description": "create a new student",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "createStudent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "student id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.StudentInfosWithOrders"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
        "/api/students/private": {
            "get": {
                "description": "search students with condition",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "searchStudents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search student with name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search student with telephone",
                        "name": "telephone",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search student with address",
                        "name": "address",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search student with author_id",
                        "name": "author_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search student order by column name",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "student list page size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "student list page index",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.StudentInfoList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
        "/api/subject": {
            "post": {
                "description": "create a new subject",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subject"
                ],
                "summary": "createSubject",
                "parameters": [
                    {
                        "description": "create subject request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.CreateSubjectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.IdResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
        "/api/subjects": {
            "get": {
                "description": "list all roles",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subject"
                ],
                "summary": "listRoles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "subject parent id",
                        "name": "parent_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Subject"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.Auth": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "entity.CreateRoleRequest": {
            "type": "object",
            "properties": {
                "auth_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "entity.CreateStudentRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "boolean"
                },
                "intent_subject": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "order_source_id": {
                    "type": "integer"
                },
                "telephone": {
                    "type": "string"
                }
            }
        },
        "entity.CreateSubjectRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                }
            }
        },
...
        "entity.OrderInfoDetails": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "intent_subject": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "org_name": {
                    "type": "string"
                },
                "publisher_id": {
                    "type": "integer"
                },
                "publisher_name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "student_id": {
                    "type": "integer"
                },
                "student_name": {
                    "type": "string"
                },
                "student_telephone": {
                    "type": "string"
                },
                "to_org_id": {
                    "type": "integer"
                }
            }
        },
...
        "entity.Role": {
            "type": "object",
            "properties": {
                "auth_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Auth"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "entity.StatisticGraph": {
            "type": "object",
            "properties": {
                "performances_graph": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.StatisticRecord"
                    }
                },
                "students_graph": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.StatisticRecord"
                    }
                }
            }
        },
        "entity.StatisticRecord": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "month": {
                    "type": "integer"
                },
                "value": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "entity.StudentInfo": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "authorID": {
                    "type": "integer"
                },
                "authorName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "intent_subject": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "order_source_id": {
                    "type": "integer"
                },
                "order_source_name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "telephone": {
                    "type": "string"
                }
            }
        },
        "entity.StudentInfoList": {
            "type": "object",
            "properties": {
                "students": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.StudentInfo"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "entity.StudentInfosWithOrders": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "authorID": {
                    "type": "integer"
                },
                "authorName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "intent_subject": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "order_source_id": {
                    "type": "integer"
                },
                "order_source_name": {
                    "type": "string"
                },
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.OrderInfoDetails"
                    }
                },
                "status": {
                    "type": "integer"
                },
                "telephone": {
                    "type": "string"
                }
            }
        },
...
        "entity.Subject": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent": {
                    "$ref": "#/definitions/entity.Subject"
                },
                "parent_id": {
                    "type": "integer"
                }
            }
        },
...
        "route.IdResponse": {
            "type": "object",
            "properties": {
                "err_msg": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "route.Response": {
            "type": "object",
            "properties": {
                "err_msg": {
                    "type": "string"
                }
            }
        },
...
    }
}